tree_incl_down.hh: fix a bug for checking emptiness
Fixed a bug when checking whether the language of a tuple is empty.
Previously, when there was a state in the tuple with a non-empty language, the
algorithm returned that inclusion *does not* hold. As the language of a tuple
is the cartesian product of languages of components of the tuple, this is
non-sense, as such a language is empty *iff* there is some component with the
empty language.